+++ src/components/template/layout.tsx
@@ -1,0 +1,15 @@
+import React, { ReactNode } from 'react';
+import Header from '../organism/header';
+
+type Props = {
+  children: ReactNode;
+}
+
+const Layout = ({ children }: Props) => (
+  <div className='layout'>
+    <Header />
+    { children }
+  </div>
+);
+
+export default Layout;
