--- src/pages/components/page/layout.tsx
@@ -1,14 +1,0 @@
-import react from 'react';
-import Header from '../organism/header';
-
-const Layout = props => {
-  const { children } = props;
-  return (
-    <div className='layout'>
-      <Header />
-      { children }
-    </div>
-  );
-}
-
-export default Layout;
